code/land_deed_system/requirements.txt
openai
pydantic
sqlalchemy
networkx
fastapi
uvicorn
python-dotenv
